refactor: rename cache to editor_cache
