+++ src/harlequin/editor_cache.py
